# runtime-acm/src/main/resources/openapi/openapi.yaml
...
openapi: 3.0.1
info:
  title: Api Documentation
  description: Api Documentation
  termsOfService: urn:tos
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  version: "1.0"
servers:
  - url: https://{server}
    variables:
      server:
        default: onap/acm/v3
        description: This value is assigned by the service provider
...
paths:
...
  /compositions:
    get:
      tags:
      - Automation Composition Definition
      summary: Query the commissioned automation composition definitions
        definitions
      description: Query the commissioned automation composition
        definitions, returning the automation composition details
      operationId: queryCompositionDefinitions
      parameters:
      - name: name
        in: query
        required: false
        description: Automation composition definition name. Regular expressions are supported for filtering. If
          this parameter is not specified, all automation composition definitions are returned.
        schema:
          type: string
      - name: version
        in: query
        required: false
        description: Automation composition definition version. Regular expressions are supported for filtering. If this
          parameter is not specified, all automation composition definitions that match the "name" filter are are returned.
        schema:
          type: string
      - name: X-onap-RequestId
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      responses:
        200:
          description: Serialised instance of
            [ToscaServiceTemplates](https://github.com/onap/policy-models/blob/master/models-tosca/src/main/java/org/onap/policy/models/tosca/authorative/concepts/ToscaServiceTemplates.java)
            that contains the automation composition definitions that match the requested filters.
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToscaServiceTemplates'
              example:
                externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/getAllCompositionDefinitions.json'
            application/yaml:
              schema:
                $ref: '#/components/schemas/ToscaServiceTemplates'
              example:
                externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/getAllCompositionDefinitions.yaml'
        401:
          description: Authentication Error, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
      security:
      - basicAuth: []
      x-interface info:
        api-version: 1.0.0
        last-mod-release: London
    post:
      tags:
      - Automation Composition Definition
      summary: Commissions automation composition definitions
      description: Commissions automation composition definitions, returning the UUIDs of
        automation composition definitions commissioned by this request.
      operationId: createCompositionDefinitions
      parameters:
      - name: X-onap-RequestId
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      requestBody:
        description: Serialised instance of
            [ToscaServiceTemplate](https://github.com/onap/policy-models/blob/master/models-tosca/src/main/java/org/onap/policy/models/tosca/authorative/concepts/ToscaServiceTemplate.java)
            containing the automation composition definitions to be commissioned.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToscaServiceTemplate'
            example:
              externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/postCompositionDefinitions.json'
          application/yaml:
            schema:
              $ref: '#/components/schemas/ToscaServiceTemplate'
            example:
              externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/postCompositionDefinitions.yaml'
        required: true
      responses:
        201:
          description: Serialised instance of
            [CommissioningResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/commissioning/CommissioningResponse.java)
            containing the UUIDs of automation composition definitions created by this request
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommissioningResponse'
              example:
                externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/postCommissionCompositionDefinitionsResponse.json'
            application/yaml:
              schema:
                $ref: '#/components/schemas/CommissioningResponse'
              example:
                externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/postCommissionCompositionDefinitionsResponse.yaml'
        401:
          description: Authentication Error, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        400:
          description: Bad Request, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
      security:
      - basicAuth: []
      x-interface info:
        api-version: 1.0.0
        last-mod-release: London
      x-codegen-request-body-name: body
...
  /compositions/{compositionId}/instances:
    get:
      tags:
      - Automation Composition Instance
      summary: Query details of the requested automation composition instances
      description: Query details of the requested automation composition instances for the given automation composition definition ID, returning
        details of all its automation composition instances
      operationId: queryCompositionInstances
      parameters:
      - name : compositionId
        in: path
        description: The UUID of the automation composition definition for which to return instances
        required: true
        schema:
          type: string
          format: uuid
      - name: name
        in: query
        description: Automation composition instance name. Regular expressions are supported for filtering. If
          this parameter is not specified, all automation composition instances for the specified definition are returned.
        schema:
          type: string
      - name: version
        in: query
        description: Automation composition instance version. Regular expressions are supported for filtering. If this
          parameter is not specified, all automation composition instances for the specified definition that match the "name" filter are are returned.
        schema:
          type: string
      - name: X-onap-RequestId
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      responses:
        200:
          description: Serialised instance of
            [AutomationCompositions](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/concepts/AutomationCompositions.java)
            containing a list of automation composition instances found
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationCompositions'
              example:
                externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/getCompositionInstancesResponse.json'
            application/yaml:
              schema:
                $ref: '#/components/schemas/AutomationCompositions'
              example:
                externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/getCompositionInstancesResponse.yaml'
        401:
          description: Authentication Error, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        404:
          description: The specified automation composition definition was not found, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
      security:
      - basicAuth: []
      x-interface info:
        api-version: 1.0.0
        last-mod-release: London
    post:
      tags:
      - Automation Composition Instance
      summary: Create automation composition instance
      description: Creates automation composition instance that use the sepcified automation composition definition. The ID of the created
        automation composition instance is returned.
      operationId: createCompositionInstance
      parameters:
      - name : compositionId
        in: path
        description: The UUID of the automation composition definition on which to create instance
        required: true
        schema:
          type: string
          format: uuid
      - name: X-onap-RequestId
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      requestBody:
        description: Serialised instance of
          [AutomationComposition](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/concepts/AutomationComposition.java)
          containing a automation composition instance to create
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/AutomationComposition'
            example:
              externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/postCompositionInstance.json'
          application/yaml:
            schema:
                $ref: '#/components/schemas/AutomationComposition'
            example:
              externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/postCompositionInstance.yaml'
        required: true
      responses:
        201:
          description: Serialised instance of
            [InstantiationResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/instantiation/InstantiationResponse.java)
            containing the UUID of the created automation composition instance
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstantiationResponse'
              example:
                externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/postCompositionInstancesResponse.json'
            application/yaml:
              schema:
                $ref: '#/components/schemas/InstantiationResponse'
              example:
                externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/postCompositionInstancesResponse.yaml'
        401:
          description: Authentication Error, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        404:
          description: The specified automation composition definition was not found, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        400:
          description: Bad Request, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
      security:
      - basicAuth: []
      x-interface info:
        api-version: 1.0.0
        last-mod-release: London
      x-codegen-request-body-name: body
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    ToscaServiceTemplates:
      title: ToscaServiceTemplates
      type: object
    ToscaServiceTemplate:
      title: ToscaServiceTemplate
      type: object
    AutomationComposition:
      title: AutomationComposition
      type: object
    AutomationCompositions:
      title: AutomationCompositions
      type: object
    SimpleResponse:
      title: SimpleResponse
      type: object
    CommissioningResponse:
      title: CommissioningResponse
      type: object
...
    InstantiationResponse:
      title: InstantiationResponse
      type: object